Project Support
-add pubspec.yaml
-flutter lib/main
-dart bin
-https://api.example.com/data is deprecated
