+++ pubspec.yaml
@@ -1,0 +1,34 @@
+name: isolate_examples
+version: 0.1.0
+
+description:
+  Isolate examples
+
+homepage: https://github.com/jwinarske/isolates.dart
+
+environment:
+  sdk: '>=3.5.0 <4.0.0'
+
+platforms:
+  linux:
+
+dependencies:
+  args: ^2.5.0
+  compute: ^1.0.2
+  ffi: ^2.1.3
+  flutter:
+    sdk: flutter
+  http: ^1.2.2
+
+dev_dependencies:
+  lints: ^4.0.0
+  test: ^1.25.8
+  test_cov: ^1.0.1
+
+executables:
+  loading-compute: loading_compute
+  loading-isolate: loading_isolate
+  loading-main-thread: loading_main_thread
+  search-compute: search_compute
+  search-isolate: search_isolate
+  search-main-thread: search_main_thread
